% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intCPUE-package.R
\docType{package}
\name{intCPUE-package}
\alias{intCPUE}
\alias{intCPUE-package}
\title{intCPUE package}
\description{
Integrated TMB-based framework for CPUE standardization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RujiaBi/intCPUE}
}

}
\author{
\strong{Maintainer}: Rujia Bi \email{rbi@iattc.org}

}
