% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{intCPUE}
\alias{intCPUE}
\title{Fit intCPUE model (encounter + positive)}
\usage{
intCPUE(
  formula,
  data,
  DLL = "intCPUE",
  vessel_effect = c("on", "off"),
  q_diffs_time = c("on", "off"),
  q_diffs_spatial = c("on", "off"),
  control = list(eval.max = 1e+05, iter.max = 1e+05),
  ...,
  silent = FALSE
)
}
\arguments{
\item{formula}{A model formula with optional mgcv::s() smooth terms.
(For now, formula is used only to build smooth objects; other fixed effects
are not yet parsed in this draft because your cpp currently only uses
vessel/year/flag + spatial fields + smooths.)}

\item{data}{A data.frame with required columns.}

\item{DLL}{Name of compiled TMB DLL (without extension).}

\item{vessel_effect}{"on" or "off". If "off", vessel RE is fixed at 0 via map.}

\item{q_diffs_spatial}{"on" or "off". If "off", flag_s_1/2 spatial fields are fixed at 0 via map.}

\item{control}{List. e.g. control = list(eval.max=1e5, iter.max=1e5, profile=FALSE)}

\item{...}{Passed to make_data() (mesh/scale settings).}
}
\value{
A list of class "intCPUE" containing obj, opt, rep, prep, maps, etc.
}
\description{
Fit intCPUE model (encounter + positive)
}
\author{
Rujia Bi \email{rbi@iattc.org}
}
